Create Code to deal with exception handling
This is code for exception handling that will catch the exception and display a user friendly text to the user that they can understand. This code deals with adding two of the same items to the database. This code can be used as a template for the rest of the exception handling.